+++ src/Enums/TransactionType.php
@@ -8,6 +8,16 @@ enum TransactionType: string
 {
     case Purchase = 'Purchase';
     case Authorize = 'Authorize';
+    case Authorization = 'Authorization';
+    case Verification = 'Verification';
+    case OffsitePurchase = 'OffsitePurchase';
+    case OffsiteAuthorization = 'OffsiteAuthorization';
+    case ExportPaymentMethods = 'ExportPaymentMethods';
+    case ReplacePaymentMethod = 'ReplacePaymentMethod';
+    case ContactCardHolder = 'ContactCardHolder';
+    case NoUpdate = 'NoUpdate';
+    case Inquiry = 'Inquiry';
+    case ScaAuthentication = 'Sca::Authentication';
     case Capture = 'Capture';
     case Void = 'Void';
     case Credit = 'Credit';